Modularized and updated the boundary condition for extraction
