+++ backend/config.py
@@ -1,0 +1,10 @@
+import os
+
+UPLOAD_FOLDER = 'uploads'
+TEMPLATE_FOLDER = 'resources/json_templates'
+ALLOWED_EXTENSIONS = {'pdf'}
+
+if not os.path.exists(UPLOAD_FOLDER):
+    os.makedirs(UPLOAD_FOLDER)
+if not os.path.exists(TEMPLATE_FOLDER):
+    os.makedirs(TEMPLATE_FOLDER)
