+++ backend/modules/utils.py
@@ -1,0 +1,2 @@
+# utils.py
+# Placeholder for future utility functions
