+++ backend/requirements.txt
@@ -1,6 +1,8 @@
-flask
-flask-cors
-pytesseract
-pdf2image
-Pillow
-pandas
+Flask==2.0.2
+Flask-Cors==3.0.10
+pytesseract==0.3.8
+pdf2image==1.16.0
+Werkzeug==2.0.2
+opencv-python==4.5.4.60
+numpy==1.21.4
+python-dateutil
