+++ resources/json_templates/Test Template1.json
@@ -1,68 +1,126 @@
-[
-  {
-    "index": 1,
-    "keyword": "VAT REG NO",
-    "name": "VAT REG NO",
-    "separator": ":"
-  },
-  {
-    "index": 5,
-    "keyword": "Sales Invoice No",
-    "name": "Sales Invoice No",
-    "separator": " "
-  },
-  {
-    "index": 9,
-    "keyword": "Invoice Time",
-    "name": "Invoice Time",
-    "separator": " "
-  },
-  {
-    "index": 5,
-    "keyword": "Date of Invoice",
-    "name": "Date of Invoice",
-    "separator": " "
-  },
-  {
-    "index": 1,
-    "keyword": "Sales Person",
-    "name": "Sales Person",
-    "separator": "Sales Person : "
-  },
-  {
-    "index": 1,
-    "keyword": "Sales Type",
-    "name": "Sales Type",
-    "separator": ":"
-  },
-  {
-    "index": 1,
-    "keyword": "Comptroller",
-    "name": "Comptroller",
-    "separator": ":"
-  },
-  {
-    "index": 1,
-    "keyword": "Billed To",
-    "name": "Billed To",
-    "separator": ":"
-  },
-  {
-    "index": 1,
-    "keyword": "Telephone",
-    "name": "Telephone",
-    "separator": ":"
-  },
-  {
-    "index": 1,
-    "keyword": "TRN",
-    "name": "TRN",
-    "separator": ":"
-  },
-  {
-    "index": 9,
-    "keyword": "Total",
-    "name": "Total",
-    "separator": "|"
-  }
-]
+{
+  "name": "Test Template",
+  "fields": [
+    {
+      "index": "1",
+      "keyword": "VAT REG NO",
+      "name": "VAT REG NO",
+      "separator": ":",
+      "boundaries": {
+        "left": "",
+        "right": ""
+      },
+      "data_type": "number"
+    },
+    {
+      "index": "5",
+      "keyword": "Sales Invoice No",
+      "name": "Sales Invoice No",
+      "separator": " ",
+      "boundaries": {
+        "left": "",
+        "right": "Invoice Time"
+      },
+      "data_type": "text"
+    },
+    {
+      "index": "9",
+      "keyword": "Invoice Time",
+      "name": "Invoice Time",
+      "separator": " ",
+      "boundaries": {
+        "left": "Sales Invoice No",
+        "right": ""
+      },
+      "data_type": "time"
+    },
+    {
+      "index": "5",
+      "keyword": "Date of Invoice",
+      "name": "Date of Invoice",
+      "separator": " ",
+      "boundaries": {
+        "left": "",
+        "right": "Sales Person"
+      },
+      "data_type": "date"
+    },
+    {
+      "index": "1",
+      "keyword": "Sales Person",
+      "name": "Sales Person",
+      "separator": "Sales Person : ",
+      "boundaries": {
+        "left": "Date of Invoice",
+        "right": ""
+      },
+      "data_type": "text"
+    },
+    {
+      "index": "1",
+      "keyword": "Sales Type",
+      "name": "Sales Type",
+      "separator": ":",
+      "boundaries": {
+        "left": "",
+        "right": "Comptroller"
+      },
+      "data_type": "text"
+    },
+    {
+      "index": "1",
+      "keyword": "Comptroller",
+      "name": "Comptroller",
+      "separator": ":",
+      "boundaries": {
+        "left": "Sales Type",
+        "right": ""
+      },
+      "data_type": "text"
+    },
+    {
+      "index": "1",
+      "keyword": "Billed To",
+      "name": "Billed To",
+      "separator": ":",
+      "boundaries": {
+        "left": "",
+        "right": ""
+      },
+      "data_type": "text"
+    },
+    {
+      "index": "1",
+      "keyword": "Telephone",
+      "name": "Telephone",
+      "separator": ":",
+      "boundaries": {
+        "left": "",
+        "right": ""
+      },
+      "data_type": "number"
+    },
+    {
+      "index": "1",
+      "keyword": "TRN",
+      "name": "TRN",
+      "separator": ":",
+      "boundaries": {
+        "left": "",
+        "right": ""
+      },
+      "data_type": "number"
+    },
+    {
+      "index": "9",
+      "keyword": "Total",
+      "name": "Total",
+      "separator": "|",
+      "boundaries": {
+        "left": "",
+        "right": ""
+      },
+      "data_type": "number"
+    }
+  ]
+}
